requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
anthropic==0.49.0
python-dotenv==1.0.1
playwright==1.50.0
aiofiles==23.2.1
pydantic==2.11.1
